Rename to scorer.py
